requirement.txt
flask
requests
beautifulsoup4
tldextract
googlesearch-python